# swagger.json
{
  "openapi": "3.0.0",
  "info": { "title": "URL Shortener API", "version": "1.0.0" },
  "servers": [
    { "url": "http://localhost:4000", "description": "Development server" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": { "type": "apiKey", "in": "header", "name": "x-api-key" }
    }
  },
  "security": [],
  "paths": {
...
    "/api/links": {
      "post": {
        "summary": "Create a short link",
        "security": [{ "apiKey": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "destination": {
                    "type": "string",
                    "description": "The URL to shorten.",
                    "example": "https://example.com/long-url"
                  },
                  "slug": {
                    "type": "string",
                    "description": "Optional custom slug for the short URL.",
                    "example": "my-custom-slug"
                  },
                  "title": {
                    "type": "string",
                    "description": "Optional title for the link.",
                    "example": "My Awesome Product Page"
                  },
                  "meta": {
                    "type": "object",
                    "description": "Optional metadata for the link."
                  },
                  "password": {
                    "type": "string",
                    "description": "Optional password to protect the link.",
                    "example": "mypassword"
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Optional expiration date for the link.",
                    "example": "2023-12-31T23:59:59Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Short link created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "slug": { "type": "string" },
                    "shortUrl": { "type": "string" },
                    "id": { "type": "integer" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid request or link creation failed." },
          "401": { "description": "Unauthorized, API key missing or invalid." }
        }
      },
      "get": {
        "summary": "List all short links for the authenticated API key",
        "security": [{ "apiKey": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "description": "Page number for pagination.",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "description": "Number of links per page.",
            "schema": { "type": "integer", "default": 10 }
          }
        ],
        "responses": {
          "200": {
            "description": "List of short links.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "links": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer" },
                          "slug": { "type": "string" },
                          "destination": { "type": "string" },
                          "title": { "type": "string" },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "expiresAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "total": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized, API key missing or invalid." }
        }
      }
    },
    "/api/links/{slug}/stats": {
      "get": {
        "summary": "Get statistics for a short link",
        "security": [{ "apiKey": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "description": "The slug of the short link.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Link statistics retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "slug": { "type": "string" },
                    "destination": { "type": "string" },
                    "clicksCount": { "type": "integer" },
                    "clicks": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized, API key missing or invalid." },
          "404": { "description": "Link not found." }
        }
      }
    },
    "/api/links/{id}": {
      "put": {
        "summary": "Update a short link",
        "security": [{ "apiKey": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The ID of the short link to update.",
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "destination": {
                    "type": "string",
                    "description": "The URL to shorten.",
                    "example": "https://example.com/long-url"
                  },
                  "slug": {
                    "type": "string",
                    "description": "Optional custom slug for the short URL.",
                    "example": "my-custom-slug"
                  },
                  "title": {
                    "type": "string",
                    "description": "Optional title for the link.",
                    "example": "My Awesome Product Page"
                  },
                  "meta": {
                    "type": "object",
                    "description": "Optional metadata for the link."
                  },
                  "password": {
                    "type": "string",
                    "description": "Optional password to protect the link.",
                    "example": "mypassword"
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Optional expiration date for the link.",
                    "example": "2023-12-31T23:59:59Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Short link updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer" },
                    "slug": { "type": "string" },
                    "shortUrl": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid request or link update failed." },
          "401": { "description": "Unauthorized, API key missing or invalid." },
          "404": { "description": "Link not found." }
        }
      },
      "delete": {
        "summary": "Delete a short link",
        "security": [{ "apiKey": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The ID of the short link to delete.",
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Short link deleted successfully." },
          "401": { "description": "Unauthorized, API key missing or invalid." },
          "404": { "description": "Link not found." }
        }
      }
    },
...
  },
...
}